Refactor runners further
