+++ src/olmo_eval/evals/tasks/__init__.py
@@ -11,6 +11,7 @@ from .core import (
     Task,
     TaskConfig,
     clear_registry,
+    get_base_task_name,
     get_task,
     list_regimes,
     list_tasks,
@@ -29,6 +30,7 @@ __all__ = [
     "Task",
     "TaskConfig",
     "clear_registry",
+    "get_base_task_name",
     "get_task",
     "list_regimes",
     "list_tasks",
