+++ src/olmo_eval/runners/agent/__init__.py
@@ -1,0 +1,5 @@
+"""Agent evaluation runner for multi-turn agent tasks."""
+
+from olmo_eval.runners.agent.runner import AgentEvalRunner
+
+__all__ = ["AgentEvalRunner"]
