+++ tests/test_runner.py
@@ -5,8 +5,7 @@ import pytest
 # Import to ensure tasks and suites are registered
 import olmo_eval.evals  # noqa: F401
 import olmo_eval.evals.tasks  # noqa: F401
-from olmo_eval.runners import EvalRunner
-from olmo_eval.runners.synchronous import ValidationError
+from olmo_eval.runners import EvalRunner, ValidationError
 
 
 class TestEvalRunnerValidation:
